Add 4.21

+++ Osa_4/blogilista/controllers/blogs.js
@@ -52,8 +52,26 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'invalid token' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).end()
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(403).json({ error: 'you don\'t have permission to delete this blog' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
+
   } catch (error) {
     next(error)
   }
